backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
requests
python-dateutil
gunicorn
whitenoise
dj-database-url
psycopg2-binary